refactor: 코드 컨벤션 및 코드 수정, 웹툰 좋아요 기능 구현
