--- src/main/java/com/project/webtoonzoa/dto/WebtoonRequestDto.java
@@ -1,15 +1,0 @@
-package com.project.webtoonzoa.dto;
-
-import com.project.webtoonzoa.entity.Enum.Category;
-import com.project.webtoonzoa.entity.Enum.Day;
-import lombok.Getter;
-
-@Getter
-public class WebtoonRequestDto {
-
-    private String title;
-    private String description;
-    private Category category;
-    private String author;
-    private Day day;
-}
